탄막과의 충돌 시 화면에 이펙트 효과 구현 Stage2에 해당하는 패턴 구현

+++ Project 3DSimpleAvoid/Bullet.h
@@ -17,6 +17,6 @@ private:
 
 	float mSpeed;
 	float mAccel;
-	int mCountTime;
+	float mCountTime;
 };
